requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dill==0.3.9
firebase-admin==6.7.0
Flask==3.1.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.1
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpcio==1.72.0rc1
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
msgpack==1.1.0
narwhals==1.34.1
ntlm-auth==1.5.0
numpy==1.23.5
packaging==24.2
pandas==2.2.3
pbr==6.1.1
pillow==11.1.0
ply==3.11
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydeck==0.9.1
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
qiskit==0.43.2
qiskit-aer==0.12.1
qiskit-ibmq-provider==0.20.2
qiskit-terra==0.24.1
referencing==0.36.2
requests==2.32.3
requests-ntlm==1.1.0
rpds-py==0.24.0
rsa==4.9
rustworkx==0.16.0
scipy==1.15.2
six==1.17.0
smmap==5.0.2
stevedore==5.4.1
streamlit==1.44.1
symengine==0.9.2
sympy==1.13.3
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
